CryptoList now has an additional dictionary to resolve the symbol search more quickly.
split up the market values of the binance failed in certain situation
where the market starts or ends with the coin symbol but there are
remaining characters in the market name. The solution was to remove the
symbol name of the coin in the for loop from the market, get the crypto
from the list of cryptos and check if the market name is equal to the
concatenation of both crypto names.

In the export function of the DeltaParser, the currency has to be
converted for certain types of cryptos, e.g. IOTA. Therefore, a new
map was created where all cryptos have to be defined which needs special
attention.