Update Converter
Renamed the _PARSER variable to PARSER.

The format command line argument now specificies the format of the
**output** file and not the format of the **input** file as in the
previous versions. The `Converter` now tries to parse the specified
file with all available parsers to generate a list of transactions.

A new variable EXPORTER is introduced to define the available export
formats (e.g. all parser with implemented export method).

We now use the logging module for giving some state information on runtime.